added a test and make sure to give error page if needed

+++ src/main/java/com/luv2code/springmvc/controller/GradebookController.java
@@ -50,6 +50,9 @@ public class GradebookController {
 
     @GetMapping("/delete/student/{id}")
     public String deleteStudent(@PathVariable int id, Model m) {
+        if(id <= 0 ){
+            return "error";// cant use "redirect:/error" because it will return redirect status (300~399)
+        }
 
         studentService.deleteStudent(id);
         Iterable<CollegeStudent> students = studentService.getGradeBook();
